Update test data creation
- Refactor test data creation to multiple files
- Create mote reservations, and some "busy" reservation units
- Add permissions to all pre-created users to access admin panel
